:truck: Rename #10: ArtLetter -> ArtBoard로 네이밍 수정
