+++ app/src/main/java/com/umc/edison/domain/usecase/artboard/GetArtBoardUseCase.kt
@@ -1,0 +1,4 @@
+package com.umc.edison.domain.usecase.artboard
+
+class GetArtBoardUseCase {
+}
