--- app/src/main/java/com/umc/edison/domain/usecase/artletter/GetArtLetterUseCase.kt
@@ -1,4 +1,0 @@
-package com.umc.edison.domain.usecase.artletter
-
-class GetArtLetterUseCase {
-}
